refactor: move runner capability tool types to runner-contract (stage 2)
Continues decoupling the optional runner feature from core. Moves two more
behavior-free contract types out of runner into homeboy-runner-contract:
- `RunnerRequiredTool` (data + pure constructors new/homeboy/git/id)
- `RunnerToolCapabilityRequirement`

`RunnerRequiredTool::remediation()` consulted core's runner tool registry, so
it can't live on the contract type (an inherent impl can't span crates). It's
now a `required_tool_remediation(&RunnerRequiredTool)` free function in core's
runner/capabilities.rs; the two call sites are updated.

Both types are re-exported from `runner::capabilities` so internal/CLI call
sites resolve unchanged. External-core consumers (upgrade/runners,
rig/capabilities) are repointed to `homeboy_runner_contract::`, severing those
reverse edges.

Deferred: RunnerCapabilityPreflight stays in core this stage — it has a
`From<PreparedLabRunnerCapability>` that couples it to a core type (orphan
rule), so it moves when the lab-capability cluster does.

Part of the staged runner extraction plan (wiki: homeboy-runner-crate-
extraction-plan).

Verified: cargo check clean for -p homeboy-runner-contract, -p homeboy-core
--lib, -p homeboy-cli --lib.
