+++ crates/homeboy-core/src/upgrade/runners/source_checkout.rs
@@ -5,11 +5,11 @@ use crate::runner;
 use crate::runner::Runner;
 use crate::runner::RunnerCapabilityPreflight;
 use crate::runner::RunnerExecOptions;
-use crate::runner::RunnerRequiredTool;
 use crate::runner::RunnerWorkspaceSyncMode;
 use crate::runner::RunnerWorkspaceSyncOptions;
 use crate::Result;
 use homeboy_runner_contract::RunnerKind;
+use homeboy_runner_contract::RunnerRequiredTool;
 use std::path::Path;
 use std::process::Command;
